Fix UnhashableTypeError message; remove InternalHashError
