+++ lib/tests/streamlit/caching/memo_test.py
@@ -23,7 +23,6 @@ import streamlit as st
 from streamlit import StreamlitAPIException, file_util
 from streamlit.caching.cache_errors import CacheError
 from streamlit.caching.memo_decorator import _cache_info
-from tests import testutil
 
 
 class MemoTest(unittest.TestCase):
@@ -257,44 +256,3 @@ class MemoPersistTest(unittest.TestCase):
             "Unsupported persist option 'yesplz'. Valid values are 'disk' or None.",
             str(e.exception),
         )
-
-
-class MemoErrorsTest(testutil.DeltaGeneratorTestCase):
-    """Make sure user-visible error messages look correct.
-
-    These errors are a little annoying to test, but they're important! So we
-    are testing them word-for-word as much as possible. Even though this
-    *feels* like an antipattern, it isn't: we're making sure the codepaths
-    that pull useful debug info from the code are working.
-    """
-
-    def test_st_warning_text(self):
-        @st.experimental_memo
-        def st_warning_text_func():
-            st.markdown("hi")
-
-        st_warning_text_func()
-
-        el = self.get_delta_from_queue(-2).new_element
-        self.assertEqual("CachedStFunctionWarning", el.exception.type)
-
-        expected_message = """
-Your script uses `st.markdown()` or `st.write()` to write to your Streamlit app
-from within some cached code at `st_warning_text_func()`. This code will only be
-called when we detect a cache "miss", which can lead to unexpected results.
-
-How to fix this:
-* Move the `st.markdown()` or `st.write()` call outside `st_warning_text_func()`.
-* Or, if you know what you're doing, use `@st.experimental_memo(suppress_st_warning=True)`
-to suppress the warning.
-        """
-        self.assertEqual(
-            testutil.normalize_md(expected_message),
-            testutil.normalize_md(el.exception.message),
-        )
-        self.assertNotEqual(len(el.exception.stack_trace), 0)
-        self.assertEqual(el.exception.message_is_markdown, True)
-        self.assertEqual(el.exception.is_warning, True)
-
-        el = self.get_delta_from_queue(-1).new_element
-        self.assertEqual(el.markdown.body, "hi")
